requirements.txt
absl-py==2.3.1
altgraph==0.17.4
argcomplete==3.6.3
astunparse==1.6.3
attrs==25.4.0
auto-py-to-exe==2.48.1
beautifulsoup4==4.14.2
black==25.11.0
bottle==0.13.4
bottle-websocket==0.2.9
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
EbookLib==0.20
Eel==0.18.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
flatbuffers==25.9.23
fonttools==4.60.1
fpdf2==2.8.4
future==1.0.0
gast==0.4.0
gevent==25.9.1
gevent-websocket==0.10.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflate64==1.0.4
joblib==1.5.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
ml-dtypes==0.2.0
multivolumefile==0.2.3
mypy_extensions==1.1.0
numpy==1.25.2
oauthlib==3.3.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pefile==2023.2.7
pillow==11.3.0
pipx==1.8.0
platformdirs==4.4.0
playwright==1.56.0
protobuf==4.25.8
psutil==7.1.3
py7zr==1.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycryptodomex==3.23.0
pyee==13.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
pywin32-ctypes==0.2.3
pyzstd==0.18.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tomli==2.3.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
userpath==1.9.2
webdriver-manager==4.0.1
Werkzeug==3.1.4
wrapt==1.14.2
wsproto==1.2.0
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
